Refactoring honor privilege
