+++ accesscontroltool-bundle/src/main/java/biz/netcentric/cq/tools/actool/honor/HonorPrivilegeServiceImpl.java
@@ -1,11 +1,9 @@
-package biz.netcentric.cq.tools.actool.aceservice.impl;
+package biz.netcentric.cq.tools.actool.honor;
 
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedSet;
-import java.util.TreeSet;
 
 import javax.jcr.Node;
 import javax.jcr.RepositoryException;
@@ -21,14 +19,14 @@ import org.apache.sling.jcr.api.SlingRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import biz.netcentric.cq.tools.actool.aceservice.impl.model.PathACL;
+import biz.netcentric.cq.tools.actool.aceservice.impl.AceServiceImpl;
 import biz.netcentric.cq.tools.actool.helper.AccessControlUtils;
 import biz.netcentric.cq.tools.actool.installationhistory.AcHistoryService;
 import biz.netcentric.cq.tools.actool.installationhistory.AcInstallationHistoryPojo;
 
 @Service
 @Component
-public class PrivilegeFacadeImpl implements PrivilegeFacade {
+public class HonorPrivilegeServiceImpl implements HonorPrivilegeService {
 
 	private static final Logger LOG = LoggerFactory.getLogger(AceServiceImpl.class);
 	
@@ -38,34 +36,13 @@ public class PrivilegeFacadeImpl implements PrivilegeFacade {
 	@Reference
 	private SlingRepository repository;
 
-	public void installHonoredPrivileges(Map<String, SortedSet<PathACL>> honoredACL) 
+	public void installHonoredPrivileges(Set<PathACL> honoredACL) 
 			throws RepositoryException {
-		Session session = null;
-		try {
-			for (Map.Entry<String, SortedSet<PathACL>> entry : honoredACL.entrySet()) {
-				this.resetPermissions(entry.getKey(), entry.getValue());
-			}
-		} finally {
-			if (session != null) {
-				session.logout();
-			}
-		}
-	}
-
-	private void resetPermissions(String group, SortedSet<PathACL> pathACLs) throws RepositoryException {
 		Session session = null;
 		
 		 try {
 			session = repository.loginAdministrative(null);		
-			for (PathACL pathACL : pathACLs) {
-				// Purge all permissions for the group on this path:
-				/*
-				this.removeExistingACL(session, group, pathACL.getPath());
-				org.apache.jackrabbit.commons.jackrabbit.authorization.AccessControlUtils.clear(
-						session.getNode(pathACL.getPath()), group);
-				LOG.info("Clearing " + pathACL.getPath() + " for group " + group);	
-				*/
-				// Apply saved permissions for this path
+			for (PathACL pathACL : honoredACL) {
 				AccessControlUtils.applyAccessControlList(session, pathACL.getPath(), pathACL.getAcl());
 			}
 			session.save();
@@ -76,24 +53,7 @@ public class PrivilegeFacadeImpl implements PrivilegeFacade {
 			}
 		}
 	}
-	
-	private void removeExistingACL(Session session, String group, String path) throws RepositoryException {
-	
-		Node node = session.getNode(path);
-		
-		org.apache.jackrabbit.commons.jackrabbit.authorization.AccessControlUtils.clear(node, group);
-		
-			
-		for (Node subNode : JcrUtils.getChildNodes(session.getNode(path))) {
-			
-			String subNodePath = subNode.getPath();
-			// Ensure that node is not meta-data, such as rep:policy
-			if (subNodePath.indexOf(':') < 0) {
-				LOG.debug("Clearing " + subNode.getPath() + " for group " + group);	
-			}
-		}	
-	}
-	
+
 
 	/**
 	 * Returns the ACLs for each path and group in honoredPrivileges, including
@@ -106,14 +66,14 @@ public class PrivilegeFacadeImpl implements PrivilegeFacade {
 	 * @throws RepositoryException
 	 *             in case of error
 	 */
-	public Map<String, SortedSet<PathACL>> getHonoredACL(Map<String, SortedSet<String>> pathsByGroup, AcInstallationHistoryPojo history)
+	public Set<PathACL> getHonoredACL(Map<String, SortedSet<String>> pathsByGroup, AcInstallationHistoryPojo history)
 			throws RepositoryException {
 
-		HashMap<String, SortedSet<PathACL>> result = new HashMap<>();
+		Set<PathACL> result = new HashSet<>();
 		Session session = repository.loginAdministrative(null);
 
 		for (Map.Entry<String, SortedSet<String>> entry : pathsByGroup.entrySet()) {
-			SortedSet<PathACL> acls = new TreeSet<>();
+			Set<PathACL> acls = new HashSet<>();
 			for (String path : entry.getValue()) {
 				if (("/").equals(path)) { 
 					String message = "Honor privilege on root folder ignored.";
@@ -123,7 +83,7 @@ public class PrivilegeFacadeImpl implements PrivilegeFacade {
 					acls.addAll(this.findRecursiveACLs(session, entry.getKey(), path));
 				}
 			}
-			result.put(entry.getKey(), acls);
+			result.addAll(acls);
 		}
 
 		return result;
